+++ alpha/lib/enums/BatchJobType.php
@@ -57,4 +57,5 @@ interface BatchJobType extends BaseEnum
 	const STORAGE_PERIODIC_EXPORT = 53;
 	const STORAGE_PERIODIC_PURGE = 54;
 	const STORAGE_PERIODIC_DELETE_LOCAL = 55;
+	const REACH_JOB_CLEANER = 55;
 }